Added configuration options for recording indicator (now w/helptext)
Menu choices allow selection of OFF, Placement in the Info Bar, Placement on the Screen and display of the raw frame buffers for debugging. Each level shows progressively more information (but is correspondingly more intrusive on the display). Thanks to Stevefal and 1% for the feedback and A1ex for review.

hg-changeset: ce693ecb140b681c065ae23dbf95335b645710cc
hg-branch: unified