Merging d6-dumper, fix major break with bmp_vram_info scoping.
The fix is a bit of a bodge.  Here's the problem: modules bring
in dryos.h, which brings in lots of model specific stuff.
Specifically, the split of the boot code into Digic < 6 and
Digic >= 6 caused problems in vram.h, because modules don't
define a Digic version and the #ifdefs get confused.

This needs a cleaner fix (currently I hide bmp_vram_info definition
so it's only inside the CONFIG_DIGIC blocks, and luckily enough
the module code doesn't ever refer to it so we're okay.  Not
good enough long term.

Also:
 - a bunch of minor fixes for slightly bad merge
 - some small changes to fix noisy compile warnings