Change indexed LUT to make clrscr() nicer
clrscr() does a bmp_fill() with index 0.  This is labelled
as COLOR_EMPTY and commented as "transparent", but I think the
comment is a lie, it's expected to overwrite the existing content.

Maybe this is some difference between pre-D6 cams, but for now,
on post D6, forcing this black looks better.

Also removes some unnecessary debug printfs.