Arkanoid module crash fixes, D >= 6 support
There were some definite bugs that could lead to null
pointer derefs and crashes.  There was an integer overflow
bug that couldn't be hit on old cams, but is possible to
hit on new cams that currently lack bfnt support and we
have a hackish workaround.

I don't understand how this code could run without crashing
on old cams, due to accessing NULL prev and next pointers
in the elem list.  So I may be missing something.

This still has lots of debug printfs which should be removed.
But, it still wants testing - not sure all crashes are found.
Before, some crashes only happened after a minute or more
of playing (stochastic?  Heap exhaustion?).