modules: remove weird delete target
This target existed for use with checkdep.py, which I removed previously.
That would make a build "fail" after the .mo successfully built, in order
to delete it.  Yes, this was a weird system: the idea was if the deps
couldn't be satisfied, the module would get deleted and hence not included
in the per cam zip, and the next cam would rebuild it (which might allow the
deps to be met, since different cams export different symbols).

Now we do the dep checks per cam not per module, and don't need to delete
modules, we copy only the modules we need into the zip, not all that exist.